+++ metro.config.js
@@ -1,4 +1,3 @@
-// metro.config.js
 const { getDefaultConfig } = require("expo/metro-config");
 
 const defaultConfig = getDefaultConfig(__dirname);
